+++ src/test/java/com/trycloud/step_definitions/US10_FilesModuleStepDefs.java
@@ -7,10 +7,29 @@ import com.trycloud.utilities.Driver;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
+import org.junit.Assert;
 
-public class US10_FileUploadStepDefs {
+public class US10_FilesModuleStepDefs {
 
+    SettingsPage settingsPage = new SettingsPage();
     FileUploadPage fileUploadPage = new FileUploadPage();
+
+    @When("user click Settings on the left bottom corner")
+    public void userClickSettingsOnTheLeftBottomCorner() {
+        settingsPage.filesModule.click();
+        settingsPage.settingsButton.click();
+    }
+
+    @Then("verify user can click any buttons")
+    public void verifyUserCanClickAnyButtons() {
+        BrowserUtils.waitFor(2);
+        boolean beforeSelect = settingsPage.checkBoxesValue.get(2).isSelected();
+        settingsPage.checkBoxesShow.get(2).click();
+        BrowserUtils.waitFor(2);
+        boolean afterSelect = settingsPage.checkBoxesValue.get(2).isSelected();
+
+        Assert.assertTrue(!beforeSelect==afterSelect);
+    }
 
     @When("user check the current storage usage")
     public void userCheckTheCurrentStorageUsage() {
